% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkeymap-package.R
\docType{data}
\name{cities}
\alias{cities}
\title{Cities of Turkey}
\format{An object of class \code{data.frame} with 81 rows and 14 columns.}
\usage{
cities
}
\description{
A data set containing the cities of Turkey
}
\keyword{datasets}
