% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkeymap-package.R
\docType{data}
\name{cvsh}
\alias{cvsh}
\title{GIS sources for Turkey}
\format{An object of class \code{data.frame} with 70 rows and 5 columns.}
\usage{
cvsh
}
\description{
A data set containing some GIS sources for Turkey
}
\keyword{datasets}
