% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkeymap-package.R
\docType{data}
\name{districts}
\alias{districts}
\title{Districts of Turkey}
\format{An object of class \code{data.frame} with 996 rows and 24 columns.}
\usage{
districts
}
\description{
A data set containing the districts of Turkey
}
\keyword{datasets}
