% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turkeymap-package.R
\docType{package}
\name{turkeymap-package}
\alias{turkeymap}
\alias{turkeymap-package}
\title{turkeymap: Provides city and district data for Turkey}
\description{
Provides city and district data for
    Turkey.
}
\keyword{internal}
